input_method: Ignore unsupported hints.
When the text-input client uses a version of the interface newer than the input-method client uses, the compositor will ignore unsupported pieces of the protocol.

It's not feasible to refuse the newer hints, because the input-method client may change at runtime, so instead hints are silently dropped as not to confuse the input method.